+++ src/tm_devices/device_manager.py
@@ -19,6 +19,7 @@ from typing_extensions import TypeVar
 from tm_devices.components import DMConfigParser
 from tm_devices.drivers.api.rest_api.margin_testers.margin_tester import MarginTester
 from tm_devices.drivers.api.rest_api.rest_api_device import RESTAPIDevice
+from tm_devices.drivers.device import Device
 from tm_devices.drivers.pi.data_acquisition_systems.data_acquisition_system import (
     DataAcquisitionSystem,
 )
@@ -40,6 +41,7 @@ from tm_devices.helpers import (
     DeviceTypes,
     DMConfigOptions,
     get_model_series,
+    PACKAGE_NAME,
     print_with_timestamp,
     PYVISA_PY_BACKEND,
     SerialConfig,
@@ -62,37 +64,37 @@ if TYPE_CHECKING:
     from pyvisa.resources import MessageBasedResource
     from typing_extensions import Self
 
-    from tm_devices.drivers.device import Device
-
 ####################################################################################################
 # Type Aliases
 ####################################################################################################
-AFGAlias = TypeVar("AFGAlias", bound=AFG, default=AFG)  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+AFGAlias = TypeVar("AFGAlias", bound=AFG, default=AFG)
 """An alias to a specific Arbitrary Function Generator Python driver."""
-AWGAlias = TypeVar("AWGAlias", bound=AWG, default=AWG)  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+AWGAlias = TypeVar("AWGAlias", bound=AWG, default=AWG)
 """An alias to a specific Arbitrary Waveform Generator Python driver."""
-DataAcquisitionSystemAlias = TypeVar(  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+DataAcquisitionSystemAlias = TypeVar(
     "DataAcquisitionSystemAlias", bound=DataAcquisitionSystem, default=DataAcquisitionSystem
 )
 """An alias to a specific Data Acquisition System Python driver."""
-DigitalMultimeterAlias = TypeVar(  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+DigitalMultimeterAlias = TypeVar(
     "DigitalMultimeterAlias", bound=DigitalMultimeter, default=DigitalMultimeter
 )
 """An alias to a specific Digital Multimeter Python driver."""
-ScopeAlias = TypeVar("ScopeAlias", bound=Scope, default=Scope)  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+ScopeAlias = TypeVar("ScopeAlias", bound=Scope, default=Scope)
 """An alias to a specific Scope driver."""
-MarginTesterAlias = TypeVar("MarginTesterAlias", bound=MarginTester, default=MarginTester)  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+MarginTesterAlias = TypeVar("MarginTesterAlias", bound=MarginTester, default=MarginTester)
 """An alias to a specific Margin Tester Python driver."""
-PowerSupplyUnitAlias = TypeVar(  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+PowerSupplyUnitAlias = TypeVar(
     "PowerSupplyUnitAlias", bound=PowerSupplyUnit, default=PowerSupplyUnit
 )
 """An alias to a specific Power Supply Unit Python driver."""
-SourceMeasureUnitAlias = TypeVar(  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+SourceMeasureUnitAlias = TypeVar(
     "SourceMeasureUnitAlias", bound=SourceMeasureUnit, default=SourceMeasureUnit
 )
 """An alias to a specific Source Measure Unit Python driver."""
-SystemsSwitchAlias = TypeVar("SystemsSwitchAlias", bound=SystemsSwitch, default=SystemsSwitch)  # pylint: disable=unexpected-keyword-arg,useless-suppression  # TODO: remove pylint disable statement
+SystemsSwitchAlias = TypeVar("SystemsSwitchAlias", bound=SystemsSwitch, default=SystemsSwitch)
 """An alias to a specific Systems Switch Python driver."""
+UnsupportedDeviceAlias = TypeVar("UnsupportedDeviceAlias", bound=Device, default=Device)
+"""An alias to a custom device driver for an unsupported device type."""
 
 
 ####################################################################################################
@@ -618,6 +620,55 @@ class DeviceManager(metaclass=Singleton):
             ),
         )
 
+    def add_unsupported_device(
+        self,
+        address: str,
+        *,
+        alias: Optional[str] = None,
+        connection_type: Optional[str] = None,
+        port: Optional[int] = None,
+        gpib_board_number: Optional[int] = None,
+    ) -> UnsupportedDeviceAlias:
+        """Add a custom device to the DeviceManager that is not an officially supported device type.
+
+        !!! warning
+            This should not be used unless absolutely necessary.
+
+        Args:
+            address: The address of the device, either an IP address or hostname. If the connection
+                     type is ``"USB"`` then the address must be specified as ``"<model>-<serial>"``.
+            alias: An optional alias to use to refer to the device. If no alias is provided,
+                   the device type and number can be used to access the device instead.
+            connection_type: The type of connection to use for VISA, defaults to TCPIP, not needed
+                when the address is a visa resource expression since the connection type is parsed
+                from the address string.
+            port: The port to use when creating a socket connection.
+            gpib_board_number: The GPIB board number (also referred to as a controller) to be used
+                when making a GPIB connection (defaults to 0).
+
+        Returns:
+            The custom device driver.
+        """
+        self.__protect_access()
+        warnings.warn(
+            f"An unsupported device type is being added to the {self.__class__.__name__}. "
+            f"Not all functionality will be available in the device driver. "
+            f"Please consider contributing to {PACKAGE_NAME} to implement official "
+            f"support for this device type.",
+            stacklevel=2,
+        )
+        return cast(
+            UnsupportedDeviceAlias,
+            self._add_device(
+                device_type=DeviceTypes.UNKNOWN.value,
+                address=address,
+                alias=alias,
+                connection_type=connection_type,
+                port=port,
+                gpib_board_number=gpib_board_number,
+            ),
+        )
+
     def cleanup_all_devices(self) -> None:
         """Cleanup and reset all devices."""
         self.__protect_access()
@@ -818,7 +869,11 @@ class DeviceManager(metaclass=Singleton):
             message = f"{device_name} was not found in the device driver dictionary."
             raise LookupError(message) from error
         # double check that the device is the correct type
-        if device_type is not None and device.device_type != device_type.upper():
+        if (
+            device_type is not None
+            and device.device_type != device_type.upper()
+            and device.config_entry.device_type != DeviceTypes.UNKNOWN
+        ):
             message = (
                 f'A device of type "{device_type}" was specified to be accessed, '
                 f'but the accessed device type is actually of type "{device.device_type}".'
@@ -1241,7 +1296,10 @@ class DeviceManager(metaclass=Singleton):
             self.__devices[device_config.alias] = new_device
 
         # double check created device is correct type
-        if new_device.device_type != new_device.config_entry.device_type.value:
+        if (
+            new_device.device_type != new_device.config_entry.device_type.value
+            and new_device.config_entry.device_type != DeviceTypes.UNKNOWN
+        ):
             self.remove_device(
                 alias=new_device.config_entry.alias,
                 device_type=new_device.config_entry.device_type.value,