+++ src/tm_devices/drivers/pi/digital_multimeters/digital_multimeter.py
@@ -1,7 +1,5 @@
 """Base Digital Multimeter (DMM) device driver module."""
 
-import inspect
-
 from abc import ABC
 
 from tm_devices.drivers.pi.tsp_device import TSPDevice
@@ -24,14 +22,3 @@ class DigitalMultimeter(TSPDevice, ABC):
     ################################################################################################
     # Private Methods
     ################################################################################################
-    def _reboot(self) -> None:
-        """Perform the actual rebooting code.
-
-        Raises:
-            NotImplementedError: Indicates the current driver has not implemented this method.
-        """
-        # TODO: implement
-        raise NotImplementedError(
-            f"``.{inspect.currentframe().f_code.co_name}()``"  # pyright: ignore[reportOptionalMemberAccess]
-            f" is not yet implemented for the {self.__class__.__name__} driver"
-        )