+++ src/tm_devices/drivers/pi/source_measure_units/smu24xx/smu24xx_interactive.py
@@ -74,9 +74,3 @@ class SMU24xxInteractive(SourceMeasureUnit, ABC):
     ################################################################################################
     # Private Methods
     ################################################################################################
-    def _reboot(self) -> None:
-        """Perform the actual rebooting code."""
-        # TODO: implement
-        raise NotImplementedError(
-            f"``.reboot()`` is not yet implemented for the {self.__class__.__name__} driver"
-        )